升级debezium版本（1.2.1.Final升级到1.4.0.Final），增加SqlServer CDC支持

+++ flink-connector-debezium/src/main/java/com/alibaba/ververica/cdc/debezium/table/RowDataDebeziumDeserializeSchema.java
@@ -230,7 +230,7 @@ public final class RowDataDebeziumDeserializeSchema implements DebeziumDeseriali
 
 	private long convertToLong(Object dbzObj, Schema schema) {
 		if (dbzObj instanceof Integer) {
-			return (long) dbzObj;
+			return ((Integer) dbzObj).longValue();
 		} else if (dbzObj instanceof Long) {
 			return (long) dbzObj;
 		} else {
